GL_UNSIGNED_SHORT index buffers

+++ source/ref_gl/r_local.h
@@ -32,7 +32,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 typedef struct mempool_s mempool_t;
 typedef struct cinematics_s cinematics_t;
 
-typedef unsigned int elem_t;
+typedef unsigned short elem_t;
 
 typedef vec_t instancePoint_t[8]; // quaternion for rotation + xyz pos + uniform scale
